backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
supabase==2.3.4
python-dotenv==1.0.1
pydantic[email]==2.6.3
python-multipart==0.0.9
httpx==0.27.0
